Revise README, include simulator in samples
I revised the README to be easier to follow for beginners. I have also included the simulator controller and support files in the sample projects along with the FTC keystore files. The properties file is now an example so that the main one is not accidentally pushed into a repository. See README for details.